Rename: clubController, clubConverter 패키지 생성 후 관련 파일 옮김
