+++ src/main/java/com/partnerd/converter/clubConverter/ClubMembershipRequestConverter.java
@@ -1,15 +1,9 @@
-package com.partnerd.converter;
+package com.partnerd.converter.clubConverter;
 
-import com.partnerd.domain.Category;
-import com.partnerd.domain.Club;
-import com.partnerd.domain.Member;
 import com.partnerd.domain.mapping.ClubMembershipRequest;
 import com.partnerd.web.dto.clubDTO.*;
-import com.partnerd.web.dto.memberDTO.MemberResponseDTO;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 public class ClubMembershipRequestConverter {
