+++ src/main/java/com/partnerd/web/controller/clubController/ClubMembershipRequestRestController.java
@@ -1,14 +1,12 @@
-package com.partnerd.web.controller;
+package com.partnerd.web.controller.clubController;
 
 import com.partnerd.apiPaylaod.ApiResponse;
 import com.partnerd.config.security.JwtTokenProvider;
-import com.partnerd.converter.ClubMembershipRequestConverter;
+import com.partnerd.converter.clubConverter.ClubMembershipRequestConverter;
 import com.partnerd.domain.mapping.ClubMembershipRequest;
 import com.partnerd.service.clubMembershipRequestService.ClubMembershipRequestService;
 import com.partnerd.web.dto.clubDTO.*;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
